+++ source/validity.h
@@ -6,12 +6,12 @@ struct ValidityChecks {
 	bool range = false;
 
 	bool lengths = false;
-	uint32_t maxPlugins = 1000;
-	uint32_t maxStringLength = 16384; // 16k strings
+	unsigned int maxPlugins = 1000;
+	unsigned int maxStringLength = 16384; // 16k strings
 	
 	bool filterOnlyWorking = false;
 	
-	ValidityChecks(size_t level) {
+	ValidityChecks(unsigned int level) {
 		// basic range/type checks
 		if (level >= 10) {
 			range = true;
